Move UAV controllers from competition.ign to team.ign
The intention here is that the implementation of UAV controllers is the responsibility of each team. We provide an example here that launches the roll-pitch-yawrate-thrust controller from RotorS. Teams can choose other types of controllers, such as the Lee Position controller from RotorS, but such controllers might need additional sensor data to function properly. Ultimately, the interface for actuating the UAVs in simulation should be mav_msgs/Actuators, which is available on the `<robot_name>/command/motor_speed` topic.

--HG--
branch : uav_controllers_in_team